+++ src/components/Home/Intro11.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import '../styles/Home.scss';
 import Intro11Data from '../LottieAnimations/Intro11Animation';
-import LottieControl from '../shared/LottieController';
+import LottieControl from '../shared/LottieControl';
 
 function Intro11(): JSX.Element {
   return (
