+++ src/components/Home/Intro12.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import '../styles/Home.scss';
 import Intro12Data from '../LottieAnimations/Intro12Animation';
-import LottieControl from '../shared/LottieController';
+import LottieControl from '../shared/LottieControl';
 
 function Intro12(): JSX.Element {
   return (
