+++ src/components/Home/Intro9.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import '../styles/Home.scss';
 import Intro9Data from '../LottieAnimations/Intro9Animation';
-import LottieControl from '../shared/LottieController';
+import LottieControl from '../shared/LottieControl';
 
 function Intro9(): JSX.Element {
   return (
