+++ src/components/shared/LottieControl.tsx
